LayoutTests/inspector/timeline/timeline-network-resource-expected.txt

CONSOLE MESSAGE: line 7: Script evaluated
Tests the Timeline API instrumentation of a network resource load

Script resource loaded

ResourceSendRequest Properties:
{
    startTime : <number>
    data : {
        identifier : <number>
        url : <string>
        requestMethod : "GET"
    }
    type : "ResourceSendRequest"
    usedHeapSize : <number>
    totalHeapSize : <number>
}

ResourceReceiveResponse Properties:
{
    startTime : <number>
    data : {
        identifier : <number>
        statusCode : 0
        mimeType : <string>
        url : <string>
    }
    children : <object>
    endTime : <number>
    type : "ResourceReceiveResponse"
    usedHeapSize : <number>
    totalHeapSize : <number>
}

ResourceFinish Properties:
{
    startTime : <number>
    data : {
        identifier : <number>
        didFail : false
        url : <string>
    }
    type : "ResourceFinish"
    usedHeapSize : <number>
    totalHeapSize : <number>
}
